+++ ui/src/hooks/useLogVirtualizer.ts
